requirements.txt
black==22.6
elasticsearch==8.4.1
flake8==4.0
isort==5.10.1
pre-commit==2.15.0
psycopg2-binary==2.9.3
pydantic==v1.10.2
python-dotenv==0.20
